Changing from double to float data type for better accuracy. Comparison Test Class improvements

+++ src/test/java/com/testvagrant/models/CityWeather.java
@@ -7,9 +7,9 @@ public class CityWeather {
 	
 	private String cityName;
 	private String condition;
-	private float tempCelcius;
-	private float tempFarnhit;
-	private float humidity;
+	private double tempCelcius;
+	private double tempFarnhit;
+	private int humidity;
 	
 
 }
